+++ src/constants/NotificationType.ts
@@ -1,0 +1,5 @@
+
+
+export const NOTIFICATION_TYPE = {
+  requiredFieldValidation: "requiredFieldValidation",
+}