+++ src/panelWebView/components/Fields/Toggle.tsx
@@ -1,41 +1,46 @@
 import * as React from 'react';
-import { RocketIcon } from '../Icons/RocketIcon';
+import { useEffect, useMemo } from 'react';
+import { BaseFieldProps } from '../../../models';
 import { ToggleIcon } from '../Icons/ToggleIcon';
-import { VsLabel } from '../VscodeComponents';
+import { FieldTitle } from './FieldTitle';
+import { RequiredMessage } from './RequiredMessage';
 
-export interface IToggleProps {
-  label: string;
-  checked: boolean;
+export interface IToggleProps extends BaseFieldProps<boolean> {
   onChanged: (checked: boolean) => void;
 }
 
-export const Toggle: React.FunctionComponent<IToggleProps> = ({label, checked, onChanged}: React.PropsWithChildren<IToggleProps>) => {
-  const [ isChecked, setIsChecked ] = React.useState(checked);
+export const Toggle: React.FunctionComponent<IToggleProps> = ({label, value, required, onChanged}: React.PropsWithChildren<IToggleProps>) => {
+  const [ isChecked, setIsChecked ] = React.useState(value);
   
   const onChange = (event: React.ChangeEvent<HTMLInputElement>) => {
     setIsChecked(!isChecked);
     onChanged(!isChecked);
   };
 
-  React.useEffect(() => {
-    if (isChecked !== checked) {
-      setIsChecked(checked);
+  const showRequiredState = useMemo(() => {
+    return required && value === null;
+  }, [required, value]);
+
+  useEffect(() => {
+    if (isChecked !== value) {
+      setIsChecked(value);
     }
-  }, [ checked ]);
+  }, [ value ]);
 
   return (
     <div className={`metadata_field`}>
-      <VsLabel>
-        <div className={`metadata_field__label`}>
-          <ToggleIcon /> <span style={{ lineHeight: "16px"}}>{label}</span>
-        </div>
-      </VsLabel>
+      <FieldTitle 
+        label={label}
+        icon={<ToggleIcon />}
+        required={required} />
 
     
       <label className="field__toggle">
-        <input type="checkbox" checked={isChecked} onChange={onChange} />
+        <input type="checkbox" checked={value === null ? false : value} onChange={onChange} />
         <span className="field__toggle__slider"></span>
       </label>
+      
+      <RequiredMessage name={label.toLowerCase()} show={showRequiredState} />
     </div>
   );
 };